Refactor HFP provider
