+++ DCT.TraineeTasks.BluetoothChargeCheck.sln
@@ -12,15 +12,13 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BluetoothChargeCheck.Models", "BluetoothChargeCheck.Models\BluetoothChargeCheck.Models.csproj", "{344D0DDA-74B0-4443-AF4E-CE6A55A0123B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BluetoothChargeCheck.Models", "BluetoothChargeCheck.Models\BluetoothChargeCheck.Models.csproj", "{344D0DDA-74B0-4443-AF4E-CE6A55A0123B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BluetoothChargeCheck.ViewModels", "BluetoothChargeCheck.ViewModels\BluetoothChargeCheck.ViewModels.csproj", "{7D85E163-EBD0-4533-8658-00EE482CADD9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BluetoothChargeCheck.ViewModels", "BluetoothChargeCheck.ViewModels\BluetoothChargeCheck.ViewModels.csproj", "{7D85E163-EBD0-4533-8658-00EE482CADD9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BluetoothChargeCheck.Resources", "BluetoothChargeCheck.Resources\BluetoothChargeCheck.Resources.csproj", "{630107FB-2FA9-4AA8-AF73-7FCBA70AE230}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BluetoothChargeCheck.Resources", "BluetoothChargeCheck.Resources\BluetoothChargeCheck.Resources.csproj", "{630107FB-2FA9-4AA8-AF73-7FCBA70AE230}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BluetoothChargeCheck.Helpers", "BluetoothChargeCheck.Helpers\BluetoothChargeCheck.Helpers.csproj", "{70EDDA87-F207-4AFA-9139-6FCC21B597E9}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BluetoothChargeCheck.Core", "BluetoothChargeCheck.Core\BluetoothChargeCheck.Core.csproj", "{7AFA07E0-9D11-41AC-B061-0E20180713A1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BluetoothChargeCheck.Core", "BluetoothChargeCheck.Core\BluetoothChargeCheck.Core.csproj", "{7AFA07E0-9D11-41AC-B061-0E20180713A1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -146,26 +144,6 @@ Global
 		{630107FB-2FA9-4AA8-AF73-7FCBA70AE230}.Release|x64.Build.0 = Release|Any CPU
 		{630107FB-2FA9-4AA8-AF73-7FCBA70AE230}.Release|x86.ActiveCfg = Release|Any CPU
 		{630107FB-2FA9-4AA8-AF73-7FCBA70AE230}.Release|x86.Build.0 = Release|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Debug|ARM.Build.0 = Debug|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Debug|x64.Build.0 = Debug|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Debug|x86.Build.0 = Debug|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Release|ARM.ActiveCfg = Release|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Release|ARM.Build.0 = Release|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Release|ARM64.Build.0 = Release|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Release|x64.ActiveCfg = Release|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Release|x64.Build.0 = Release|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Release|x86.ActiveCfg = Release|Any CPU
-		{70EDDA87-F207-4AFA-9139-6FCC21B597E9}.Release|x86.Build.0 = Release|Any CPU
 		{7AFA07E0-9D11-41AC-B061-0E20180713A1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7AFA07E0-9D11-41AC-B061-0E20180713A1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7AFA07E0-9D11-41AC-B061-0E20180713A1}.Debug|ARM.ActiveCfg = Debug|Any CPU
